Seperate AsyncOpenTaskHandle and NotificationHandle from realm_core.dart
