+++ packages/realm_dart/lib/src/list.dart
@@ -12,8 +12,8 @@ import 'collections.dart';
 import 'native/collection_changes_handle.dart';
 import 'native/handle_base.dart';
 import 'native/list_handle.dart';
+import 'native/notification_token_handle.dart';
 import 'native/object_handle.dart';
-import 'native/realm_core.dart';
 import 'realm_class.dart';
 import 'realm_object.dart';
 import 'results.dart';
