+++ packages/realm_dart/lib/src/map.dart
@@ -12,8 +12,8 @@ import 'collections.dart';
 import 'native/handle_base.dart';
 import 'native/map_changes_handle.dart';
 import 'native/map_handle.dart';
+import 'native/notification_token_handle.dart';
 import 'native/object_handle.dart';
-import 'native/realm_core.dart';
 import 'realm_object.dart';
 import 'realm_class.dart';
 import 'results.dart';
