+++ packages/realm_dart/lib/src/native/set_handle.dart
@@ -7,6 +7,7 @@ import 'package:ffi/ffi.dart';
 
 import '../realm_dart.dart'; // TODO: remove this import
 import 'error_handling.dart';
+import 'notification_token_handle.dart';
 import 'query_handle.dart';
 import 'realm_bindings.dart';
 import 'realm_core.dart'; // TODO: Remove this import
