+++ packages/realm_dart/lib/src/realm_object.dart
@@ -10,8 +10,8 @@ import 'package:realm_common/realm_common.dart';
 import 'configuration.dart';
 import 'list.dart';
 import 'native/handle_base.dart';
+import 'native/notification_token_handle.dart';
 import 'native/object_handle.dart';
-import 'native/realm_core.dart';
 import 'native/realm_library.dart';
 import 'realm_class.dart';
 import 'results.dart';
