+++ packages/realm_dart/lib/src/results.dart
@@ -9,8 +9,8 @@ import 'package:cancellation_token/cancellation_token.dart';
 import 'collections.dart';
 import 'native/collection_changes_handle.dart';
 import 'native/handle_base.dart';
+import 'native/notification_token_handle.dart';
 import 'native/object_handle.dart';
-import 'native/realm_core.dart';
 import 'native/results_handle.dart';
 import 'realm_class.dart';
 import 'realm_object.dart';
