+++ packages/realm_dart/lib/src/set.dart
@@ -10,8 +10,8 @@ import 'package:collection/collection.dart' as collection;
 
 import 'native/collection_changes_handle.dart';
 import 'native/handle_base.dart';
+import 'native/notification_token_handle.dart';
 import 'native/object_handle.dart';
-import 'native/realm_core.dart';
 import 'native/set_handle.dart';
 import 'realm_class.dart';
 import 'realm_object.dart';
